py-scouter/pyproject.toml
[project]
name = "scouter-ml"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
license = "MIT"
version = "0.25.0"
description = ""
authors = [
    {name = "Thorrester", email = "<support@demmlai.com>"},
    {name = "russellkemmit", email = "<support@demmlai.com>"}
]
readme = "README.md"
dependencies = []

[project.optional-dependencies]
opentelemetry = [
    "opentelemetry-sdk>=1.36.0, <1.39.0",
    "opentelemetry-exporter-otlp-proto-common",
    "opentelemetry-instrumentation",
]

[dependency-groups]
dev = [
    "maturin >= 1.4.0, < 2.0.0",
    "pytest >= 7.0.0",
    "pytest-cov >= 5.0.0",
    "ruff >= 0.1.0, < 1.0.0",
    "black >= 24.3.0, < 25.0.0",
    "pylint >= 4.0.0, < 5.0.0",
    "pytest-lazy-fixture >= 0.6.3, < 1.0.0",
    "psutil >= 5.9.8, < 6.0.0",
    "isort >= 5.13.2, < 6.0.0",
    "pandas-stubs >= 1.5.3.230214, < 3.0.0",
    "fastapi >= 0.115.0, < 1.0.0",
    "uvicorn >= 0.31.0, < 1.0.0",
    "pip>=24.3.1",
    'pandas>=1.5.3, <3.0.0; python_version < "3.10"',
    'pandas>=2.2.2; python_version >= "3.10"',
    'numpy>1.24.4; python_version >= "3.10"',
    'pydantic>=2.7.1, <3.0.0',
    'polars>=1.0.0, <2.0.0',
    'pyarrow>=22.0.0',
    "httpx>=0.28.1",
    "pytest-benchmark>=5.2.3",
    "google-adk>=1.27.2",
    "pytest-asyncio>=1.3.0",
    "crewai>=1.6.1",
    "ty>=0.0.29",
    "openinference-instrumentation-crewai>=1.1.2",
    "openai-agents>=0.14.4",
    "opentelemetry-instrumentation-openai-agents-v2>=0.1.0",
]
docs = [
    "mkdocs-material >= 9.5.4, < 10.0.0",
    "mkdocstrings[python] >= 0.26.0, < 1.0.0",
]

[tool.isort]
profile = "black"

[tool.ruff]
exclude = [
    "python/scouter/utils/comparison/*",
    "python/scouter/stubs/*",
]

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'

[tool.pylint.MASTER]
ignore-paths = ["python/scouter/utils/comparison/*"]
load-plugins = ["pylint.extensions.docparams"]
ignore-patterns = [
    "tests/*",
    "python/scouter/stubs/*",
]

[tool.pylint.messages_control]
max-line-length = 130
disable = [
    "too-few-public-methods",
    "design",
    "duplicate-code",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "too-many-nested-blocks",
    "unused-argument",
    "fixme",
    "import-outside-toplevel",
    "import-self",
    "too-many-lines",
    "invalid-name",
    "unidiomatic-typecheck",
]

[tool.flake8]
# This section is just a doc placeholder..see setup.cfg
max-complexity = 10
format = "pylint"
statistics = "True"
max-line-length = 125
# ignore non-PEP8 lints
ignore = ["E203", "W503", "W0511"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
testpaths = ["tests"]

[tool.coverage.run]
omit = [
    "tests/*",
    "python/scouter/drift/base.py"
]

[tool.ty.environment]
python-version = "3.12"

[tool.ty.src]
include = ["python/scouter", "tests", "examples"]
exclude = [
    "python/scouter/utils/comparison/*",
    "python/scouter/stubs/*",
]

[build-system]
requires = ["maturin>=1.4,<2.0"]
build-backend = "maturin"

[tool.maturin]
python-source = "python"
module-name = "scouter._scouter"

[tool.uv]
package = false
